Комит в dz11 ветку dz11-libs-nik
