+++ src/main/java/pro/sky/dz11libsnik/service/DepartmentService.java
@@ -1,43 +1,14 @@
 package pro.sky.dz11libsnik.service;
 
-
 import pro.sky.dz11libsnik.entity.Employee;
-import pro.sky.dz11libsnik.exception.EmployeeNotFoundException;
-import org.springframework.stereotype.Service;
 
-import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
 
-import static java.util.stream.Collectors.groupingBy;
-import static java.util.stream.Collectors.toList;
+public interface DepartmentService {
+    Employee getEmployeeWithMaxSalary(Integer departmentId);
 
-@Service
-public class DepartmentService {
+    Employee getEmployeeWithMinSalary(Integer departmentId);
 
-    private final EmployeeService employeeService;
-
-    public DepartmentService(EmployeeService employeeService) {
-        this.employeeService = employeeService;
-    }
-
-    public Employee getEmployeeWithMaxSalary(Integer departmentId) {
-        return employeeService.getAll().stream()
-                .filter(employee -> employee.getDepartmentId() == departmentId)
-                .max(Comparator.comparing(Employee::getSalary))
-                .orElseThrow(() -> new EmployeeNotFoundException("Сотрудник с максимальной зарплатой не найден"));
-    }
-
-    public Employee getEmployeeWithMinSalary(Integer departmentId) {
-        return employeeService.getAll().stream()
-                .filter(employee -> employee.getDepartmentId() == departmentId)
-                .min(Comparator.comparing(Employee::getSalary))
-                .orElseThrow(() -> new EmployeeNotFoundException("Сотрудник с минимальной зарплатой не найден"));
-    }
-
-    public Map<Integer, List<Employee>> getEmployeesByDepartment(Integer departmentId) {
-        return employeeService.getAll().stream()
-                .filter(e -> departmentId == null || e.getDepartmentId() == departmentId)
-                .collect(groupingBy(Employee::getDepartmentId, toList()));
-    }
+    Map<Integer, List<Employee>> getEmployeesByDepartment(Integer departmentId);
 }
