+++ src/main/java/pro/sky/dz11libsnik/service/DepartmentServiceImpl.java
@@ -1,0 +1,46 @@
+package pro.sky.dz11libsnik.service;
+
+
+import pro.sky.dz11libsnik.entity.Employee;
+import pro.sky.dz11libsnik.exception.EmployeeNotFoundException;
+import org.springframework.stereotype.Service;
+
+import java.util.Comparator;
+import java.util.List;
+import java.util.Map;
+
+import static java.util.stream.Collectors.groupingBy;
+import static java.util.stream.Collectors.toList;
+
+@Service
+public class DepartmentServiceImpl implements DepartmentService {
+
+    private final EmployeeServiceImpl employeeService;
+
+    public DepartmentServiceImpl(EmployeeServiceImpl employeeService) {
+        this.employeeService = employeeService;
+    }
+
+    @Override
+    public Employee getEmployeeWithMaxSalary(Integer departmentId) {
+        return employeeService.getAll().stream()
+                .filter(employee -> employee.getDepartmentId() == departmentId)
+                .max(Comparator.comparing(Employee::getSalary))
+                .orElseThrow(() -> new EmployeeNotFoundException("Сотрудник с максимальной зарплатой не найден"));
+    }
+
+    @Override
+    public Employee getEmployeeWithMinSalary(Integer departmentId) {
+        return employeeService.getAll().stream()
+                .filter(employee -> employee.getDepartmentId() == departmentId)
+                .min(Comparator.comparing(Employee::getSalary))
+                .orElseThrow(() -> new EmployeeNotFoundException("Сотрудник с минимальной зарплатой не найден"));
+    }
+
+    @Override
+    public Map<Integer, List<Employee>> getEmployeesByDepartment(Integer departmentId) {
+        return employeeService.getAll().stream()
+                .filter(e -> departmentId == null || e.getDepartmentId() == departmentId)
+                .collect(groupingBy(Employee::getDepartmentId, toList()));
+    }
+}
